+++ herd-server/bin/job/base/rdhd-job-base_job.js
@@ -13,7 +13,7 @@ class Job
         // Create shell object istance
         let sh = new Shell('ssh', ['-i', Config.private_key, asset.user + '@' + asset.ip, '-p', asset.wport, task], FSMFPUtils.getSpecializedLv2Message(asset.id, code), wsServer);
   
-        //  RedHert Flexible Message Formatting Protocol compliant JSON envelopes:
+        //  RedHerd Flexible Message Formatting Protocol compliant JSON envelopes:
         if (sync)
         {
             //  Sync execution