+++ herd-server/bin/job/rdhd-job-asset_monitor.js
@@ -1,6 +1,8 @@
 'use strict';
 
-const Job = require('./base/rdhd-job-base_job');
+const Model = require('../../controllers/rdhd-ctr-model_controller');
+const LinuxJob = require('./base/rdhd-job-base_linux_job');
+const WindowsJob = require('./base/rdhd-job-base_windows_job');
 
 // jobCode: "keep_alive"
 class AssetMonitorJob
@@ -9,17 +11,28 @@ class AssetMonitorJob
 
     static run(asset, sync = false, wsServer = 'http://127.0.0.1:3001')
     {
-        if (sync)
+        try
         {
-            //  Sync execution
-            // ******************************
-            Job.do(asset, AssetMonitorJob.code, "hostname", true, false, wsServer);
+            let model = new Model();
+            let assetType = model.getTypeByAssetId(asset.id).name;
+
+            switch(assetType.toUpperCase())
+            {
+                case LinuxJob.os:
+                    LinuxJob.isAlive(asset, AssetMonitorJob.code, sync, false, wsServer);
+                    break;
+                case WindowsJob.os:
+                    WindowsJob.isAlive(asset, AssetMonitorJob.code, sync, false, wsServer);
+                    break;
+                default:
+                    // Unsupported operating system
+                    console.log("Unsupported operating system");
+                    break;
+            }
         }
-        else
+        catch (e)
         {
-            //  Async execution
-            // ******************************
-            Job.do(asset, AssetMonitorJob.code, "hostname", false, false, wsServer);
+            console.log(e);
         }
     }
 }